+++ openstack/autoscaling/v1/policies/delete.go
@@ -2,7 +2,7 @@ package policies
 
 import "github.com/opentelekomcloud/gophertelekomcloud"
 
-func Delete(client *golangsdk.ServiceClient, id string) (r DeleteResult) {
-	_, r.Err = client.Delete(client.ServiceURL("scaling_policy", id), nil)
+func Delete(client *golangsdk.ServiceClient, id string) (err error) {
+	_, err = client.Delete(client.ServiceURL("scaling_policy", id), nil)
 	return
 }
